(MFW-0.5.1): Added HashFile Method
Created a hashFile method which uses try catch blocks to handle potential exceptions. In the method, the file is read using a buffered reader by getting the file path as a parameter. It's then converted into an array of bytes, and then converted into hexadecimal format. Additionally, I added a MyFileWriterTester which will be used later.